chore: memoizing transaction parsing
